+++ crabml-core/src/backends/cpu/primitives/mod.rs
@@ -1,11 +1,13 @@
 mod add;
 mod div;
 mod mul;
+mod rms_norm;
 mod rope;
 mod silu;
 
 pub use add::add_inplace;
 pub use div::div_inplace;
 pub use mul::mul_inplace;
+pub use rms_norm::rms_norm_inplace;
 pub use rope::rope_inplace;
 pub use silu::silu_inplace;
